+++ packages/util/test/component_test.js
@@ -4,7 +4,9 @@ const { join, resolve } = require('path')
 const { testProjectPath } = require('../../../test/support/paths')
 const ComponentUtil = require('../src/component')
 
-const componentsPath = resolve(testProjectPath, 'src', 'components')
+const elementsPath = resolve(testProjectPath, 'src', 'elements')
+const modulesPath = resolve(testProjectPath, 'src', 'modules')
+const componentPaths = [elementsPath, modulesPath]
 
 describe('ComponentUtil', () => {
   describe('#componentIdToPath', () => {
@@ -15,55 +17,60 @@ describe('ComponentUtil', () => {
 
   describe('#componentIdToFilePath', () => {
     it('should return component file path for component', () => {
-      const filePath = ComponentUtil.componentIdToFilePath(componentsPath, 'button')
+      const filePath = ComponentUtil.componentIdToFilePath(componentPaths, 'formfield')
 
-      assert.strictEqual(filePath, join(componentsPath, 'button', 'component.md'))
+      assert.strictEqual(filePath, join(modulesPath, 'formfield', 'component.md'))
     })
   })
 
   describe('#componentFilePathToId', () => {
     it('should return component id for component file path', () => {
-      const filePath = join(componentsPath, 'button', 'component.md')
-      const componentId = ComponentUtil.componentFilePathToId(componentsPath, filePath)
+      const filePath = join(modulesPath, 'form', 'component.md')
+      const componentId = ComponentUtil.componentFilePathToId(componentPaths, filePath)
 
-      assert.strictEqual(componentId, 'button')
+      assert.strictEqual(componentId, 'form')
     })
 
     it('should return component id for file path', () => {
-      const filePath = join(componentsPath, 'button', 'additional.pdf')
-      const componentId = ComponentUtil.componentFilePathToId(componentsPath, filePath)
+      const filePath = join(elementsPath, 'button', 'additional.pdf')
+      const componentId = ComponentUtil.componentFilePathToId(componentPaths, filePath)
 
       assert.strictEqual(componentId, 'button')
     })
 
     it('should return component id for variant file path', () => {
-      const filePath = join(componentsPath, 'button', 'variants', 'button.pug')
-      const componentId = ComponentUtil.componentFilePathToId(componentsPath, filePath)
+      const filePath = join(elementsPath, 'button', 'variants', 'button.pug')
+      const componentId = ComponentUtil.componentFilePathToId(componentPaths, filePath)
 
       assert.strictEqual(componentId, 'button')
     })
 
-    it('should return null for invalid file path', () => {
+    it('should return undefined for invalid file path', () => {
       const filePath = join(testProjectPath, 'src', 'uiengine', 'pages', 'page.md')
-      const componentId = ComponentUtil.componentFilePathToId(componentsPath, filePath)
+      const componentId = ComponentUtil.componentFilePathToId(componentPaths, filePath)
 
-      assert.strictEqual(componentId, null)
+      assert.strictEqual(componentId, undefined)
     })
   })
 
   describe('#componentPathToId', () => {
     it('should return component id for component path', () => {
-      const componentPath = join(componentsPath, 'button')
-      const componentId = ComponentUtil.componentPathToId(componentsPath, componentPath)
+      const elementPath = join(elementsPath, 'button')
+      const elementId = ComponentUtil.componentPathToId(componentPaths, elementPath)
 
-      assert.strictEqual(componentId, 'button')
+      assert.strictEqual(elementId, 'button')
+
+      const modulePath = join(elementsPath, 'form')
+      const moduleId = ComponentUtil.componentPathToId(componentPaths, modulePath)
+
+      assert.strictEqual(moduleId, 'form')
     })
 
-    it('should return null for invalid component id', () => {
+    it('should return undefined for invalid component id', () => {
       const filePath = join(testProjectPath, 'src', 'uiengine', 'pages', 'page.md')
-      const componentId = ComponentUtil.componentPathToId(componentsPath, filePath)
+      const componentId = ComponentUtil.componentPathToId(componentPaths, filePath)
 
-      assert.strictEqual(componentId, null)
+      assert.strictEqual(componentId, undefined)
     })
   })
 })