+++ packages/util/test/variant_test.js
@@ -4,28 +4,30 @@ const { join, resolve } = require('path')
 const VariantUtil = require('../src/variant')
 
 const { testProjectPath } = require('../../../test/support/paths')
-const componentsPath = resolve(testProjectPath, 'src', 'components')
+const elementsPath = resolve(testProjectPath, 'src', 'elements')
+const modulesPath = resolve(testProjectPath, 'src', 'modules')
+const componentPaths = [elementsPath, modulesPath]
 
 describe('VariantUtil', () => {
   describe('#componentIdToVariantsPath', () => {
     it('should return variants path for component id', () => {
       assert.strictEqual(
-        VariantUtil.componentIdToVariantsPath(componentsPath, 'button'),
-        join(componentsPath, 'button', 'variants'))
+        VariantUtil.componentIdToVariantsPath(componentPaths, 'form'),
+        join(modulesPath, 'form', 'variants'))
     })
   })
 
   describe('#variantFilePathToIdPrefix', () => {
     it('should return variant id for variant file path', () => {
-      const variantFilePath = join(componentsPath, 'button', 'variants', 'primary-button.pug')
-      const variantIdPrefix = VariantUtil.variantFilePathToIdPrefix(componentsPath, variantFilePath)
-      assert.strictEqual(variantIdPrefix, 'button/primary-button.pug')
+      const variantFilePath = join(elementsPath, 'input', 'variants', 'number.pug')
+      const variantIdPrefix = VariantUtil.variantFilePathToIdPrefix(componentPaths, variantFilePath)
+      assert.strictEqual(variantIdPrefix, 'input/number.pug')
     })
 
-    it('should return null for invalid variant file path', () => {
-      const variantFilePath = join(componentsPath, 'button', 'button.pug')
-      const variantId = VariantUtil.variantFilePathToIdPrefix(componentsPath, variantFilePath)
-      assert.strictEqual(variantId, null)
+    it('should return undefined for invalid variant file path', () => {
+      const variantFilePath = join(elementsPath, 'input', 'number.pug')
+      const variantId = VariantUtil.variantFilePathToIdPrefix(componentPaths, variantFilePath)
+      assert.strictEqual(variantId, undefined)
     })
   })
 
@@ -38,29 +40,29 @@ describe('VariantUtil', () => {
 
   describe('#variantFilePathToComponentId', () => {
     it('should return component id for variant file path', () => {
-      const variantFilePath = join(componentsPath, 'button', 'variants', 'primary-button.pug')
-      const componentId = VariantUtil.variantFilePathToComponentId(componentsPath, variantFilePath)
-      assert.strictEqual(componentId, 'button')
+      const variantFilePath = join(elementsPath, 'input', 'variants', 'number.pug')
+      const componentId = VariantUtil.variantFilePathToComponentId(componentPaths, variantFilePath)
+      assert.strictEqual(componentId, 'input')
     })
 
     it('should return null for invalid variant file path', () => {
-      const variantFilePath = join(componentsPath, 'button', 'button.pug')
-      const variantId = VariantUtil.variantFilePathToComponentId(componentsPath, variantFilePath)
+      const variantFilePath = join(elementsPath, 'input', 'number.pug')
+      const variantId = VariantUtil.variantFilePathToComponentId(componentPaths, variantFilePath)
       assert.strictEqual(variantId, null)
     })
   })
 
   describe('#variantIdToComponentId', () => {
     it('should return component id for variant id', () => {
-      assert.strictEqual(VariantUtil.variantIdToComponentId('button/primary-button.pug-1'), 'button')
+      assert.strictEqual(VariantUtil.variantIdToComponentId('input/number.pug-1'), 'input')
     })
   })
 
   describe('#variantIdToFilePath', () => {
     it('should return variant file path for variant id', () => {
       assert.strictEqual(
-        VariantUtil.variantIdToFilePath(componentsPath, 'input/checkbox.pug-1'),
-        join(componentsPath, 'input', 'variants', 'checkbox.pug')
+        VariantUtil.variantIdToFilePath(componentPaths, 'input/checkbox.pug-1'),
+        join(elementsPath, 'input', 'variants', 'checkbox.pug')
       )
     })
   })